server/requirements.txt
flask
flask-login
flask-wtf
flask-cors
psycopg2-binary
email-validator